feat(repo-explorer): decomposed entries view component
