+++ src/models.ts
@@ -34,3 +34,8 @@ export type BranchIdentity = {
     /** Префикс ветки */
     prefix: string;
 };
+
+/**
+ * Объект, описывающий файл или директорию
+ */
+export type GitFile = { type: string; name: string };
